+++ main.go
@@ -2,10 +2,10 @@ package main
 
 import (
 	"github.com/itsyndicate/itsb/cmd/itsb"
-	"github.com/itsyndicate/itsb/pkg/confparse"
+	// "github.com/itsyndicate/itsb/pkg/confparse"
 )
 
 func main() {
 	itsb.Execute()
-	confparse.StructTest()
+	// confparse.StructTest()
 }
